changed price to decimal format

+++ SEProject/myapp/models.py
@@ -12,41 +12,41 @@ class Category(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits = 7, decimal_places = 2)
     picture = models.ImageField(upload_to="imgs", default="")
     category = models.ForeignKey(Category, on_delete=models.CASCADE, default="No Category")
 
     def __str__(self):
         return self.name
-    
+
 class Cart(models.Model):
     id = models.UUIDField(default=uuid.uuid4, primary_key=True)
     user = models.ForeignKey(CustomUser,  on_delete=models.CASCADE)
     completed = models.BooleanField(default=False)
-    
+
     def __str__(self):
         return str(self.id)
-    
+
     @property
     def total_price(self):
         cartitems = self.cartitems.all()
         total = sum([item.price for item in cartitems])
         return total
-    
+
     @property
     def num_of_items(self):
         cartitems = self.cartitems.all()
         quantity = sum([item.quantity for item in cartitems])
         return quantity
-    
+
 class CartItem(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE, related_name='items')
     cart = models.ForeignKey(Cart, on_delete= models.CASCADE, related_name="cartitems")
     quantity = models.IntegerField(default=0)
-    
+
     def __str__(self):
         return self.product.name
-    
+
     @property
     def price(self):
         new_price = self.product.price * self.quantity
